# schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateSessionResponse": {
        "description": "`POST /sessions` response.",
        "properties": {
          "session_id": {
            "$ref": "#/components/schemas/SessionId",
            "description": "Newly created session."
          }
        },
        "required": [
          "session_id"
        ],
        "type": "object"
      },
...
      "Draft": {
        "description": "`PUT /sessions/{id}/draft` — draft synchronization (SPEC §5.A).",
        "properties": {
          "caret": {
            "description": "Caret position, in Unicode scalar values (not bytes), `0 ≤ caret ≤\ntext.chars().count()`.",
            "format": "uint32",
            "minimum": 0,
            "type": "integer"
          },
          "text": {
            "description": "Current draft text. **P0 content** — never logged.",
            "type": "string"
          }
        },
        "required": [
          "text",
          "caret"
        ],
        "type": "object"
      },
...
      "NextCharsResponse": {
        "description": "`GET /sessions/{id}/next-chars?prefix=` response — next-character\ndistribution on the warm KV (SPEC §5.A). Never triggers a full\ngeneration; feeds adaptive dwell, weighted scanning and Dasher.",
        "properties": {
          "dist": {
            "description": "Character probabilities, descending. BPE→char aggregation hub-side.",
            "items": {
              "$ref": "#/components/schemas/CharProb"
            },
            "type": "array"
          }
        },
        "required": [
          "dist"
        ],
        "type": "object"
      },
...
      "PairWindowRequest": {
        "description": "`POST /pair/window` request — opens the pairing window from the main\nUI (SPEC §2.A: pairing is only possible during an explicitly opened\nwindow; the route itself is system-scope, ADR-0005 §7).",
        "properties": {
          "scope": {
            "$ref": "#/components/schemas/Scope",
            "description": "Scope the resulting token will carry."
          }
        },
        "required": [
          "scope"
        ],
        "type": "object"
      },
      "PairWindowResponse": {
        "description": "`POST /pair/window` response — what the main screen displays.",
        "properties": {
          "code": {
            "description": "Eight-digit single-use code to read out loud or scan.",
            "type": "string"
          },
          "expires_at": {
            "description": "When the window closes (2 minutes).",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "code",
          "expires_at"
        ],
        "type": "object"
      },
...
      "Problem": {
        "description": "RFC 9457 problem document, extended with [`ErrorCode`].\n\nThe `type` URI is derived from the code (`urn:fluence:problem:<code>`):\na stable URN that does not depend on owning a domain yet; it may become\na resolvable URL later without changing the codes.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "Stable machine-readable code (the catalogue clients switch on)."
          },
          "detail": {
            "description": "Occurrence-specific explanation. **Never P0 content.**",
            "type": [
              "string",
              "null"
            ]
          },
          "instance": {
            "description": "URI identifying this occurrence (e.g. the request path).",
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "description": "HTTP status code.",
            "format": "uint16",
            "maximum": 65535,
            "minimum": 0,
            "type": "integer"
          },
          "title": {
            "description": "Short human-readable summary, stable per code (English).",
            "type": "string"
          },
          "type": {
            "description": "Stable problem-type URI (`urn:fluence:problem:<code>`).",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "type": "object"
      },
...
      "TargetMap": {
        "description": "Full declaration of a UI's selectable targets for one surface\n(`PUT /input/targets`, SPEC §4.A).\n\nWire example: `{\"surface\":\"main\",\"viewport\":{\"w\":1920,\"h\":1080},\"targets\":[…]}`",
        "properties": {
          "surface": {
            "$ref": "#/components/schemas/SurfaceId",
            "description": "Surface these targets belong to."
          },
          "targets": {
            "description": "Complete list — replaces any previous declaration for this surface.",
            "items": {
              "$ref": "#/components/schemas/Target"
            },
            "type": "array"
          },
          "viewport": {
            "$ref": "#/components/schemas/Viewport",
            "description": "Pixel size of the surface viewport (target rects are expressed in it)."
          }
        },
        "required": [
          "surface",
          "viewport",
          "targets"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "fluenceToken": {
        "description": "Per-device revocable token with a scope (SPEC §2.A)",
        "in": "header",
        "name": "X-Fluence-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Fluence",
      "url": "https://github.com/Pluenet-Killian/Fluence"
    },
    "description": "API of the Fluence hub — a local-first communication platform for people with severe motor disabilities. One session = one conversation with a warm KV-cache; suggestions stream over SSE with per-slot cancellation; input flows through the FluenceInput protocol (WebSocket topics). Authentication: per-device revocable tokens with scopes (display/control/care/system). Operations are marked `x-fluence-stability: stable` (A1 core, frozen) or `experimental` (P2 domains, may change). Errors follow RFC 9457 problem+json with a stable machine-readable code catalogue.",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "summary": "Local communication hub: input, acceleration, voice (SPEC §5.A)",
    "title": "Fluence hub API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/input/targets": {
      "put": {
        "operationId": "put_input_targets",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TargetMap"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No content"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (RFC 9457)"
          }
        },
        "security": [
          {
            "fluenceToken": []
          }
        ],
        "summary": "Declare the full target map of a surface",
        "tags": [
          "input"
        ],
        "x-fluence-scopes": [
          "control"
        ],
        "x-fluence-stability": "stable"
      }
    },
...
    "/api/v1/pair/window": {
      "post": {
        "operationId": "post_pair_window",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairWindowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairWindowResponse"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (RFC 9457)"
          }
        },
        "security": [
          {
            "fluenceToken": []
          }
        ],
        "summary": "Open the pairing window (main UI only; 2 min, single-use code)",
        "tags": [
          "pair"
        ],
        "x-fluence-scopes": [],
        "x-fluence-stability": "stable"
      }
    },
...
    "/api/v1/sessions": {
      "post": {
        "operationId": "post_sessions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSessionResponse"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (RFC 9457)"
          }
        },
        "security": [
          {
            "fluenceToken": []
          }
        ],
        "summary": "Open a conversation session (warm KV-cache)",
        "tags": [
          "sessions"
        ],
        "x-fluence-scopes": [
          "control"
        ],
        "x-fluence-stability": "stable"
      }
    },
    "/api/v1/sessions/{id}": {
      "delete": {
        "operationId": "delete_sessions_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (RFC 9457)"
          }
        },
        "security": [
          {
            "fluenceToken": []
          }
        ],
        "summary": "Close a session",
        "tags": [
          "sessions"
        ],
        "x-fluence-scopes": [
          "control"
        ],
        "x-fluence-stability": "stable"
      }
    },
    "/api/v1/sessions/{id}/draft": {
      "get": {
        "operationId": "get_sessions_id_draft",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Draft"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (RFC 9457)"
          }
        },
        "security": [
          {
            "fluenceToken": []
          }
        ],
        "summary": "Read the draft back (session resumption after a cut, SPEC §2.A)",
        "tags": [
          "sessions"
        ],
        "x-fluence-scopes": [
          "control"
        ],
        "x-fluence-stability": "stable"
      },
      "put": {
        "operationId": "put_sessions_id_draft",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Draft"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No content"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (RFC 9457)"
          }
        },
        "security": [
          {
            "fluenceToken": []
          }
        ],
        "summary": "Synchronize the draft (continuous autosave)",
        "tags": [
          "sessions"
        ],
        "x-fluence-scopes": [
          "control"
        ],
        "x-fluence-stability": "stable"
      }
    },
    "/api/v1/sessions/{id}/next-chars": {
      "get": {
        "operationId": "get_sessions_id_next_chars",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Draft prefix to condition the distribution on",
            "in": "query",
            "name": "prefix",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NextCharsResponse"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (RFC 9457)"
          }
        },
        "security": [
          {
            "fluenceToken": []
          }
        ],
        "summary": "Next-character distribution on the warm KV",
        "tags": [
          "sessions"
        ],
        "x-fluence-scopes": [
          "control"
        ],
        "x-fluence-stability": "stable"
      }
    },
    "/api/v1/sessions/{id}/suggest": {
      "post": {
        "operationId": "post_sessions_id_suggest",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestEvent"
                }
              }
            },
            "description": "Server-sent event stream. The `event:` field is the variant tag, `data:` is the JSON payload (canonical form documented by the schema)."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (RFC 9457)"
          }
        },
        "security": [
          {
            "fluenceToken": []
          }
        ],
        "summary": "Stream suggestions (SSE, per-slot cancellation)",
        "tags": [
          "sessions"
        ],
        "x-fluence-scopes": [
          "control"
        ],
        "x-fluence-stability": "stable"
      }
    },
...
    "/api/v1/system/capabilities": {
      "get": {
        "operationId": "get_system_capabilities",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitiesResponse"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (RFC 9457)"
          }
        },
        "security": [
          {
            "fluenceToken": []
          }
        ],
        "summary": "Installation tier and available features",
        "tags": [
          "system"
        ],
        "x-fluence-scopes": [
          "display",
          "control",
          "care"
        ],
        "x-fluence-stability": "stable"
      }
    },
    "/api/v1/system/emergency": {
      "post": {
        "operationId": "post_system_emergency",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmergencyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No content"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (RFC 9457)"
          }
        },
        "security": [
          {
            "fluenceToken": []
          }
        ],
        "summary": "Raise or clear the emergency alert (double-confirmed; D-7.4)",
        "tags": [
          "system"
        ],
        "x-fluence-scopes": [
          "control"
        ],
        "x-fluence-stability": "stable"
      }
    },
...
    "/api/v1/system/journal": {
      "get": {
        "operationId": "get_system_journal",
        "parameters": [
          {
            "description": "Maximum entries to return (newest first; default 100)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessJournalResponse"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (RFC 9457)"
          }
        },
        "security": [
          {
            "fluenceToken": []
          }
        ],
        "summary": "Local access journal (caregiver space; metadata only)",
        "tags": [
          "system"
        ],
        "x-fluence-scopes": [
          "care"
        ],
        "x-fluence-stability": "stable"
      }
    },
    "/api/v1/voice/speak": {
      "post": {
        "operationId": "post_voice_speak",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeakRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "audio/wav": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Streamed audio (WAV, 16-bit mono PCM — ADR-0009; Opus/Ogg for LAN/home mode is deferred to Phase 7)"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (RFC 9457)"
          }
        },
        "security": [
          {
            "fluenceToken": []
          }
        ],
        "summary": "Vocalize text (P0 scheduler priority, streamed Opus)",
        "tags": [
          "voice"
        ],
        "x-fluence-scopes": [
          "control"
        ],
        "x-fluence-stability": "stable"
      }
    },
    "/api/v1/voice/voices": {
      "get": {
        "operationId": "get_voice_voices",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoicesResponse"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (RFC 9457)"
          }
        },
        "security": [
          {
            "fluenceToken": []
          }
        ],
        "summary": "List installed voices",
        "tags": [
          "voice"
        ],
        "x-fluence-scopes": [
          "control",
          "care"
        ],
        "x-fluence-stability": "stable"
      }
    },
    "/pair": {
      "post": {
        "operationId": "post_pair",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairResponse"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (RFC 9457)"
          }
        },
        "security": [],
        "summary": "Pair a device (only while a pairing window is open)",
        "tags": [
          "pair"
        ],
        "x-fluence-stability": "stable"
      }
    },
    "/pair/info": {
      "get": {
        "operationId": "get_pair_info",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairInfo"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (RFC 9457)"
          }
        },
        "security": [],
        "summary": "Pairing screen information",
        "tags": [
          "pair"
        ],
        "x-fluence-stability": "stable"
      }
    },
    "/ws": {
      "get": {
        "operationId": "get_ws",
        "parameters": [
          {
            "description": "Comma-separated topic list (input,asr,suggest,voice,system)",
            "in": "query",
            "name": "topics",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Input protocol version (currently 1)",
            "in": "query",
            "name": "v",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Device token — browser WebSocket cannot set the X-Fluence-Token header (ADR-0004)",
            "in": "query",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "101": {
            "description": "WebSocket upgrade — frames are ServerFrame / ClientFrame (see components), first frame is system.hello"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Error (RFC 9457)"
          }
        },
        "security": [
          {
            "fluenceToken": []
          }
        ],
        "summary": "WebSocket upgrade — topics via ?topics=…&v=1 (scope-filtered)",
        "tags": [
          "ws"
        ],
        "x-fluence-scopes": [
          "display",
          "control",
          "care"
        ],
        "x-fluence-stability": "stable"
      }
    }
  },
  "servers": [
    {
      "description": "Embedded mode (loopback, no TLS)",
      "url": "http://127.0.0.1:7411"
    },
    {
      "description": "Home mode (LAN, local CA)",
      "url": "https://{host}:7411",
      "variables": {
        "host": {
          "default": "fluence.local"
        }
      }
    }
  ],
...
}